front: move XML import logic to separate file
ImportTimetableItemConfig is quite large with a lot of different
things mixed in. Let's split all XML-related logic into a separate
file.

No functional change, just moving code without any edit.
